Create pg - wrong driving table leads to bad performance
Explanation: Looking at the execution plan before tuning, the table Inspection is the driving table, however there is no filter on it => Bitmap Heap Scan on inspection. After heap scan, recheck with ramp_id, and filter on created_date to find the exact block => high cost.
In the query, most of the filters are on the table inspection_truck to reduce the number of reads. Therefore, try to create an index on this table in order to adjust the execution plan. Once, there is an index on the table inspection_truck, it can be considered as a driving table, and index scan can be used effectively (look at the execution plan after tuning)